+++ src/Node_api.cpp
@@ -780,7 +780,7 @@ std::map<std::string, vm::varptr_t> Node::read_storage_object(const addr_t& cont
 
 vnx::Variant Node::call_contract(
 		const addr_t& address, const std::string& method, const std::vector<vnx::Variant>& args,
-		const vnx::optional<addr_t>& user, const vnx::optional<std::pair<addr_t, uint80>>& deposit) const
+		const vnx::optional<addr_t>& user, const vnx::optional<std::pair<addr_t, uint128>>& deposit) const
 {
 	if(auto exec = std::dynamic_pointer_cast<const contract::Executable>(get_contract(address))) {
 		if(auto bin = std::dynamic_pointer_cast<const contract::Binary>(get_contract(exec->binary))) {
@@ -1003,7 +1003,7 @@ std::vector<offer_data_t> Node::get_recent_offers(const int32_t& limit, const vn
 }
 
 std::vector<offer_data_t> Node::get_recent_offers_for(
-		const vnx::optional<addr_t>& bid, const vnx::optional<addr_t>& ask, const uint80& min_bid, const int32_t& limit, const vnx::bool_t& state) const
+		const vnx::optional<addr_t>& bid, const vnx::optional<addr_t>& ask, const uint128& min_bid, const int32_t& limit, const vnx::bool_t& state) const
 {
 	std::vector<offer_data_t> result;
 	std::unordered_set<addr_t> bid_set;
@@ -1328,7 +1328,7 @@ std::vector<swap_entry_t> Node::get_swap_history(const addr_t& address, const in
 				if(entry.deposit) {
 					out.amount = entry.deposit->second;
 				} else if(entry.args.size() >= 2) {
-					out.amount = entry.args[1].to<uint80>();
+					out.amount = entry.args[1].to<uint128>();
 				}
 			}
 		} else if(entry.method == "rem_all_liquid") {
@@ -1352,7 +1352,7 @@ std::vector<swap_entry_t> Node::get_swap_history(const addr_t& address, const in
 	return result;
 }
 
-std::array<uint128, 2> Node::get_swap_trade_estimate(const addr_t& address, const uint32_t& i, const uint80& amount, const int32_t& num_iter) const
+std::array<uint128, 2> Node::get_swap_trade_estimate(const addr_t& address, const uint32_t& i, const uint128& amount, const int32_t& num_iter) const
 {
 	const auto info = get_swap_info(address);
