+++ src/WebAPI.cpp
@@ -1951,7 +1951,7 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 			get_context({currency}, request_id,
 				[this, request_id, args, currency](std::shared_ptr<RenderContext> context) {
 					try {
-						uint80 amount = 0;
+						uint128 amount = 0;
 						if(args["raw_mode"].to<bool>()) {
 							args["amount"].to(amount);
 						} else {
@@ -1995,7 +1995,7 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 			get_context({currency}, request_id,
 				[this, request_id, args, currency](std::shared_ptr<RenderContext> context) {
 					try {
-						std::vector<std::pair<addr_t, uint80>> amounts;
+						std::vector<std::pair<addr_t, uint128>> amounts;
 						if(args["raw_mode"].to<bool>()) {
 							args["amounts"].to(amounts);
 						} else {
@@ -2085,8 +2085,8 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 			get_context({bid_currency, ask_currency}, request_id,
 				[this, request_id, args, bid_currency, ask_currency](std::shared_ptr<RenderContext> context) {
 					try {
-						uint80 bid_amount = 0;
-						uint80 ask_amount = 0;
+						uint128 bid_amount = 0;
+						uint128 ask_amount = 0;
 						if(auto currency = context->find_currency(bid_currency)) {
 							bid_amount = to_amount(args["bid"].to<fixed128>(), currency->decimals);
 						} else {
@@ -2268,7 +2268,7 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 							}
 							const auto amount = to_amount(value, token_i->decimals);
 
-							vnx::optional<uint80> min_trade;
+							vnx::optional<uint128> min_trade;
 							if(min_value) {
 								min_trade = to_amount(*min_value, token_k->decimals);
 							}
@@ -2300,7 +2300,7 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 							const auto pool_idx = args["pool_idx"].to<uint32_t>();
 							const auto options = args["options"].to<spend_options_t>();
 
-							std::array<uint80, 2> amount = {};
+							std::array<uint128, 2> amount = {};
 							for(int i = 0; i < 2; ++i) {
 								if(const auto token = context->find_currency(info.tokens[i])) {
 									amount[i] = to_amount(value[i], token->decimals);
@@ -2470,7 +2470,7 @@ void WebAPI::http_request_async(std::shared_ptr<const vnx::addons::HttpRequest> 
 		if(iter_ask != query.end()) {
 			ask = vnx::from_string_value<addr_t>(iter_ask->second);
 		}
-		const uint80 min_bid = iter_min_bid != query.end() ? vnx::from_string<uint80>(iter_min_bid->second) : uint80();
+		const uint128 min_bid = iter_min_bid != query.end() ? vnx::from_string<uint128>(iter_min_bid->second) : uint128();
 		const uint64_t limit = iter_limit != query.end() ? vnx::from_string<int64_t>(iter_limit->second) : 100;
 		const uint64_t offset = iter_offset != query.end() ? vnx::from_string<int64_t>(iter_offset->second) : 0;
 		const bool state = iter_state != query.end() ? vnx::from_string<bool>(iter_state->second) : true;
